+++ packages/article-main-standard/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.80.15](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.80.14...@times-components/article-main-standard@3.80.15) (2023-08-30)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
 ## [3.80.14](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.80.13...@times-components/article-main-standard@3.80.14) (2023-08-29)
 
 **Note:** Version bump only for package @times-components/article-main-standard