+++ packages/article-paragraph/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.92](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.9.91...@times-components/article-paragraph@1.9.92) (2023-08-30)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
 ## [1.9.91](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.9.90...@times-components/article-paragraph@1.9.91) (2023-08-29)
 
 **Note:** Version bump only for package @times-components/article-paragraph